# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Utility Management System API",
        "description": "API documentation for the Laravel Utility Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tenants": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Get a list of tenants",
                "description": "Returns a paginated list of all tenants.",
                "operationId": "b2d6a7dfd92597bb14a6b1dd29f59a5c",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tenants retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TenantResource"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 4
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenants not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Create a new tenant",
                "description": "Creates a new tenant record.",
                "operationId": "f0081e7330ce16562f97fbe1fb6aefc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roomId",
                                    "name",
                                    "nrc",
                                    "email",
                                    "phNumber",
                                    "emergencyNo"
                                ],
                                "properties": {
                                    "roomId": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "9a7c6f2c-5b8a-4f2a-8f2c-6d8b3a0c1e9f"
                                    },
                                    "name": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "John Doe",
                                            "Jane Doe"
                                        ]
                                    },
                                    "nrc": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "12/ABC(N)123456"
                                        ]
                                    },
                                    "email": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "example": [
                                            "john.doe@example.com"
                                        ]
                                    },
                                    "phNumber": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "09123456789"
                                        ]
                                    },
                                    "emergencyNo": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "09987654321"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tenant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TenantResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Tenant creation failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tenants/{id}": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Get a single tenant",
                "description": "Returns the details of a specific tenant by their ID.",
                "operationId": "f9661d9d338c11947b7916880f02ceda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tenant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant find successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TenantResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant did not find"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Update an existing tenant",
                "description": "Updates the details of an existing tenant.",
                "operationId": "a66b6a59d9cd46deabe8d200e6be09a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tenant to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roomId",
                                    "name",
                                    "email",
                                    "phNumber",
                                    "emergencyNo"
                                ],
                                "properties": {
                                    "roomId": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "9a7c6f2c-5b8a-4f2a-8f2c-6d8b3a0c1e9f"
                                    },
                                    "name": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Johnathan Doe"
                                        ]
                                    },
                                    "nrc": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "12/DEF(N)654321"
                                        ]
                                    },
                                    "email": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "example": [
                                            "johnathan.doe@example.com"
                                        ]
                                    },
                                    "phNumber": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "09112233445"
                                        ]
                                    },
                                    "emergencyNo": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "09556677889"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tenant updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TenantResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Tenant not find"
                    },
                    "500": {
                        "description": "Tenant update failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TenantResource": {
                "title": "Tenant Resource",
                "description": "Tenant model representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "roomId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9a7c6f2c-5b8a-4f2a-8f2c-6d8b3a0c1e9f"
                    },
                    "name": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "John Doe",
                            "Jane Doe"
                        ]
                    },
                    "nrc": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "12/ABC(N)123456",
                            "12/DEF(N)654321"
                        ]
                    },
                    "email": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "email"
                        },
                        "example": [
                            "john.doe@example.com",
                            "jane.doe@example.com"
                        ]
                    },
                    "phNumber": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "09123456789",
                            "09987654321"
                        ]
                    },
                    "emergencyNo": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "09111222333",
                            "09444555666"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}